+++ fxn/compile.py
@@ -13,7 +13,7 @@ from typing import Any, Callable, Literal, ParamSpec, TypeVar, cast
 
 from .beta import (
     CoreMLInferenceMetadata, LiteRTInferenceMetadata, LlamaCppInferenceMetadata,
-    ONNXInferenceMetadata, ONNXRuntimeInferenceSessionMetadata, OpenVINOInferenceMetadata,
+    OnnxInferenceMetadata, OnnxRuntimeInferenceSessionMetadata, OpenVINOInferenceMetadata,
     QnnInferenceMetadata
 )
 from .sandbox import Sandbox
@@ -33,8 +33,8 @@ CompileMetadata = (
     CoreMLInferenceMetadata             |
     LiteRTInferenceMetadata             |
     LlamaCppInferenceMetadata           |
-    ONNXInferenceMetadata               |
-    ONNXRuntimeInferenceSessionMetadata |
+    OnnxInferenceMetadata               |
+    OnnxRuntimeInferenceSessionMetadata |
     OpenVINOInferenceMetadata           |
     QnnInferenceMetadata
 )